# openapi.yaml
openapi: 3.1.0
info:
  title: Social Neuron REST API
  version: 1.5.2
  description: |
    REST API for Social Neuron — AI-powered social media content creation, scheduling, and analytics.

    ## Authentication
    All endpoints require an API key via Bearer token:
    ```
    Authorization: Bearer snk_live_...
    ```
    Get your API key at [socialneuron.com/settings/developer](https://socialneuron.com/settings/developer).

    ## Scopes
    API keys have scopes that control access:
    - `mcp:full` — All permissions
    - `mcp:read` — Read analytics, brand, credits
    - `mcp:write` — Generate content, save plans
    - `mcp:distribute` — Schedule posts
    - `mcp:analytics` — Refresh analytics
    - `mcp:comments` — Manage comments
    - `mcp:autopilot` — Automated scheduling (Pro+)

    ## Async Operations
    Video/image generation returns `202 Accepted` with a `Location` header pointing to the job status endpoint.
    Poll `GET /v1/jobs/{id}` with the `Retry-After` interval until status is `completed` or `failed`.

    ## Rate Limits
    - Read operations: 60 requests/minute
    - Write/posting operations: 30 requests/minute
    - IP-level: 60 requests/minute (pre-auth)

    Rate-limited responses include `Retry-After` header.
  contact:
    name: Social Neuron
    url: https://socialneuron.com
    email: developers@socialneuron.com
  license:
    name: MIT
servers:
  - url: https://mcp.socialneuron.com/v1
    description: Production
  - url: http://localhost:8080/v1
...
security:
  - apiKey: []
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      description: "API key with snk_live_ prefix"
  schemas:
...
    Platform:
      type: string
      enum: [youtube, tiktok, instagram, twitter, linkedin, facebook, threads, bluesky]
...
paths:
...
  /posts:
    post:
      summary: Schedule a post
      tags: [Posts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [platforms]
              properties:
                media_url:
                  type: string
                  format: uri
                media_urls:
                  type: array
                  items:
                    type: string
                    format: uri
                media_type:
                  type: string
                  enum: [video, image, carousel]
                  default: video
                caption:
                  type: string
                title:
                  type: string
                platforms:
                  type: array
                  items:
                    $ref: "#/components/schemas/Platform"
                scheduled_at:
                  type: string
                  format: date-time
                attribution:
                  type: boolean
                  default: false
      responses:
        "201":
          description: Post scheduled
    get:
      summary: List recent posts
      tags: [Posts]
      parameters:
        - name: platform
          in: query
          schema:
            $ref: "#/components/schemas/Platform"
        - name: status
          in: query
          schema:
            type: string
        - name: days
          in: query
          schema:
            type: integer
            default: 7
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of posts
...
  /tools:
    get:
      summary: List available tools
      tags: [Tools]
      parameters:
        - name: query
          in: query
          schema:
            type: string
          description: Search tools by name/description
        - name: module
          in: query
          schema:
            type: string
          description: Filter by module
        - name: scope
          in: query
          schema:
            type: string
          description: Filter by required scope
      responses:
        "200":
          description: List of tools
  /tools/{name}:
    post:
      summary: Execute tool by name (universal proxy)
      tags: [Tools]
      description: |
        Universal tool proxy — execute any MCP tool by name.
        Pass tool parameters as JSON body. The tool's required scope
        is enforced based on your API key.
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: Tool name (e.g., generate_content, schedule_post)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Tool-specific parameters
      responses:
        "200":
          description: Tool execution result
        "202":
          description: Async job accepted (for generation tools)
        "404":
          description: Tool not found

  # ── Jobs ─────────────────────────────────────────────────────────
  /jobs/{id}:
    get:
      summary: Poll async job status
      tags: [Jobs]
      description: |
        Poll the status of an async job (video/image generation).
        Returns `Retry-After` header while job is in progress.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job status with result if completed

  # ── Account ──────────────────────────────────────────────────────
...